Refactor send_and_process functions in auth and client
