+++ tekore/_sender/__init__.py
@@ -1,4 +1,4 @@
 from .base import Sender, Request, Response
 from .concrete import SyncSender, AsyncSender
 from .extending import ExtendingSender, RetryingSender, CachingSender
-from .client import SenderConflictWarning, Client
+from .client import SenderConflictWarning, Client, send_and_process
